Move the Debhelper add-on to the dh-gstreamer1.0 package.
dh_gstscancodecs and gst-codec-info are only needed to build
GStreamer 1.0 plugin packages. So move them to a separate package.
Recommend the dh-gstreamer1.0 package from libgstreamer1.0-dev for
backward compatibility.
Note that dh-gstreamer1.0 is not yet compatible with multiarch
development because gst-codec-info can only analyze GStreamer
plugins of the same architecture.

Signed-off-by: Francois Gouget <fgouget@codeweavers.com>